+++ src/Service/HealthCheck.php
@@ -4,18 +4,22 @@ declare(strict_types=1);
 
 namespace Dbp\Relay\MonoConnectorCampusonlineBundle\Service;
 
+use Dbp\Relay\BasePersonBundle\API\PersonProviderInterface;
+use Dbp\Relay\CoreBundle\Exception\ApiError;
 use Dbp\Relay\CoreBundle\HealthCheck\CheckInterface;
 use Dbp\Relay\CoreBundle\HealthCheck\CheckOptions;
 use Dbp\Relay\CoreBundle\HealthCheck\CheckResult;
+use Dbp\Relay\CoreBundle\Rest\Options;
+use Symfony\Component\HttpFoundation\Response;
 
 class HealthCheck implements CheckInterface
 {
-    private $ldap;
     private $tuitionfee;
+    private PersonProviderInterface $personProvider;
 
-    public function __construct(LdapService $ldap, TuitionFeeService $tuitionfee)
+    public function __construct(PersonProviderInterface $personProvider, TuitionFeeService $tuitionfee)
     {
-        $this->ldap = $ldap;
+        $this->personProvider = $personProvider;
         $this->tuitionfee = $tuitionfee;
     }
 
@@ -42,11 +46,27 @@ class HealthCheck implements CheckInterface
     public function check(CheckOptions $options): array
     {
         $results = [];
-        $results[] = $this->checkMethod('Check if we can connect to the LDAP server', [$this->ldap, 'checkConnection']);
         $results[] = $this->checkMethod('Check if we can connect to the CO API', [$this->tuitionfee, 'checkConnectionNoAuth']);
         $results[] = $this->checkMethod('Check if we can authenticate with the CO API', [$this->tuitionfee, 'checkConnection']);
         $results[] = $this->checkMethod('Check if the CO API can talk to its backend', [$this->tuitionfee, 'checkBackendConnection']);
+        $results[] = $this->checkMethod('Check if the person provider is configured to deliver the title', [$this, 'checkPersonProvider']);
 
         return $results;
+    }
+
+    /**
+     * @throws \Throwable
+     */
+    public function checkPersonProvider()
+    {
+        try {
+            $personProviderOptions = [];
+            $this->personProvider->getPerson('foo', Options::requestLocalDataAttributes(
+                $personProviderOptions, [TuitionFeeService::PERSON_TITLE_LOCAL_DATA_ATTRIBUTE]));
+        } catch (\Throwable $exception) {
+            if ($exception instanceof ApiError === false || $exception->getStatusCode() !== Response::HTTP_NOT_FOUND) {
+                throw $exception;
+            }
+        }
     }
 }